Extract sanitizeHtml into a standalone module
`sanitize.ts` mixes a pure DOMPurify wrapper with React/jotai state that
transitively imports the full editor/codemirror/SQL panel tree,
including `.svg`/`.png` database icon assets added in #8528.

This breaks the marimo-lsp VS Code extension build — esbuild bundles
`transitionCell` from `core/cells/cell.ts`, which reaches `sanitize.ts`
via `outline.ts`, pulling in assets it has no loader for. This has
blocked upgrading the frontend dependency in marimo-lsp since 0.20.2.

No logic changes. `sanitizeHtml` and its DOMPurify hooks move to
`sanitize-html.ts`, `outline.ts` imports from there directly, and
`sanitize.ts` re-exports for backward compat.
